Attempt to get CI to pass without requiring arguments to cargo test
