+++ src/lib.rs
@@ -62,40 +62,3 @@ fn _pydantic_core(_py: Python, m: &PyModule) -> PyResult<()> {
     m.add_function(wrap_pyfunction!(list_all_errors, m)?)?;
     Ok(())
 }
-
-#[cfg(test)]
-mod tests {
-    // Note: these tests require the following arguments to `cargo test`:
-    // `cargo test --no-default-features --features auto-initialize`
-    use crate::SchemaSerializer;
-    use pyo3::prelude::*;
-    use pyo3::types::PyDict;
-
-    #[test]
-    fn test_build_schema_serializer() {
-        Python::with_gil(|py| {
-            let code = r#"{
-                'type': 'typed-dict',
-                'fields': {
-                    'root': {
-                        'type': 'typed-dict-field',
-                        'schema': {
-                            'type': 'definition-ref',
-                            'schema_ref': 'C-ref',
-                        },
-                    },
-                },
-                'ref': 'C-ref',
-                'serialization': {
-                    'type': 'function-wrap',
-                    'function': {
-                        'type': 'general',
-                        'function': lambda: None,
-                    },
-                },
-            }"#;
-            let schema: &PyDict = py.eval(code, None, None).unwrap().extract().unwrap();
-            SchemaSerializer::py_new(py, schema, None).unwrap();
-        })
-    }
-}
